requirements.txt
vectorbt
ccxt
pandas_ta
ta
jupyterlab